Cumulative coefficients
Source: R/cumulative-coefficient.R

These functions are designed to extract (or mimick) the cumulative coefficients usually used in additive hazards models (Aalen model) to depict (time-varying) covariate effects. For PAMMs, these are the differences between the cumulative hazard rates where all covariates except one have the identical values. For a numeric covariate of interest, we calculate \(Lambda(t|x+1) - Lambda(t|x)\). For non-numeric covariates the cumulative hazard of the reference level is substracted from the cumulative hazards eveluated at all non reference levels. Standard errors are calculated using the delta method.

# S3 method for gam
get_cumu_coef(model, data, terms, ...)

# S3 method for aalen
get_cumu_coef(model, data = NULL, terms, ci = TRUE)

Arguments

model
Object from which to extract cumulative coefficients.

data
Additional data if necessary.

terms
A character vector of variables for which the cumulative coefficient should be calculated.

...
Further arguments passed to add_cumu_hazard.

ci
Logical. Indicates if confidence intervals should be returned as well.