pammtools: Piece-wise exponential Additive Mixed Modeling tools.

`pammtools` provides functions and utilities that facilitate fitting Piece-wise Exponential Additive Mixed Models (PAMMs), including data transformation and other convenience functions for pre- and post-processing as well as plotting.

Details

The best way to get an overview of the functionality provided and how to fit PAMMs is to view the vignettes included in this package and also available at https://adibender.github.io/pammtools/. A of the vignettes content is given below:

basics: Introduction to PAMMs and basic modeling.

baseline: Shows how to estimate and visualize baseline model (without covariates) and comparison to respective Cox-PH model.

convenience: Convenience functions for post-processing and plotting PAMMs

data-transformation: Transforming data into a format suitable to fit PAMMs

frailty: Specifying "frailty" terms, i.e., random effects for PAMMs

splines: Specifying spline smooth terms for PAMMs

strata: Specifying stratified models in which each level of a grouping variable has a different baseline hazard

tdcovar: Dealing with time-dependent covariates

tveffects: Specifying time-varying effects