k-fold cross validated prediction error curve (pec)

k-fold cross validated prediction error curve (pec)

pec_cv(data, formula_pam, formula_ped = Surv(time, status) ~ ., k = 10L,
  formula_ipcw = Surv(time, status) ~ 1, cut = NULL, method = pamm, ...)

Arguments

data
The data in standard time-to-event data format which can be tansformed in PED (see split_data).

formula_pam
The formula used to fit the PAM (see pam).

formula_ped
The formula used to transform data to PED (see split_data).

k
The number of folds to be used in cross-validation.

formula_ipcw
The formula used to calculate inverse probability censoring weights (IPCW). See ipcw for details.

cut
the vector of timepoints to cut at

method
The function to which formula_pam will be passed to fit the model in the k-th fold.

...
Further arguments passed to split_data.

See also

pec, ipcw