Extract risk set information for each interval.

The columns ped_riskset, ped_events, ped_censored provide the size of the riskset at the beginning of each interval as well as the number of events and censorings that occured in the interval, respectively.

riskset_info(ped)

Arguments

ped
An object of class ped as returned by split_data.

Value

A data frame with one row for each interval in ped.

See also

int_info, sample_info

Examples

data("veteran", package="survival")
ped <- split_data(Surv(time, status)~ ., data = veteran, id = "id",
  cut = seq(0,400, by = 100))
riskset_info(ped)

#> # A tibble: 4 x 4
#>   interval  ped_riskset ped_events ped_censored
#> * <fctr>          <int>      <dbl>        <int>
#> 1 (0,100]           137      79.0             5
#> 2 (100,200]          53      26.0             3
#> 3 (200,300]          24      10.0             1
#> 4 (300,400]          13       7.00            6

(riskset_celltype <- riskset_info(group_by(ped, celltype)))

#> # A tibble: 14 x 5
#> # Groups:   celltype [4]
#>    celltype  interval  ped_riskset ped_events ped_censored
#>  * <fctr>    <fctr>          <int>      <dbl>        <dbl>
#>  1 squamous  (0,100]            35      13.0          3.00
#>  2 squamous  (100,200]          19       6.00         0   
#>  3 squamous  (200,300]          13       4.00         1.00
#>  4 squamous  (300,400]           8       3.00         5.00
#>  5 smallcell (0,100]            48      37.0          1.00
#>  6 smallcell (100,200]          10       5.00         2.00
#>  7 smallcell (200,300]           3       1.00         0   
#>  8 smallcell (300,400]           2       2.00         0   
#>  9 adeno     (0,100]            27      21.0          1.00
#> 10 adeno     (100,200]           5       5.00         0   
#> 11 large     (0,100]            27       8.00         0   
#> 12 large     (100,200]          19      10.0          1.00
#> 13 large     (200,300]           8       5.00         0   
#> 14 large     (300,400]           3       2.00         1.00

## add descriptive statistics for riskset at beginning of each interval:
# left_join(riskset_celltype,
#           group_by(ped, celltype, interval) %>% sample_info())