tidyr Verbs for ped-Objects

See tidyr documentation of the respective functions for description and examples.

# S3 method for ped
fill(data, ..., .direction = c("down", "up"),
  keep_attributes = TRUE)

Arguments

data
an object of class ped, see split_data.

...
Specification of columns to expand.

To find all unique combinations of x, y and z, including those not found in the data, supply each variable as a separate argument. To find only the combinations that occur in the data, use nest: expand(df, nesting(x, y, z)).

You can combine the two forms. For example, expand(df, nesting(school_id, student_id), date) would produce a row for every student for each date.

For factors, the full set of levels (not just those that appear in the data) are used. For continuous variables, you may need to fill in values that don't appear in the data: to do so use expressions like year = 2010:2020 or year = full_seq(year).

Length-zero (empty) elements are automatically dropped.

.direction
Direction in which to fill missing values. Currently either "down" (the default) or "up".

keep_attributes
conserve attributes? defaults to TRUE.

Value

A modified ped object.