Transform data to Piece-wise Exponential Data (PED)

This is the general data transformation function provided by the pammtools package. Two main applications must be differentiated: (1) Transformation of standard time-to-event data. (2) Transformation of time-to-event data with time-dependent covariates (TDC). For the latter, the type of effect of the TDC one wants to estimate is also important for the data transformation step, i.e. concurrent or cumulative. In any case, the data transformation is specified by a two sided formula. In case of TDCs, the right-hand-side of the formula can contain specials concurrent and cumulative. See the vignette on data transformation for details.

as_ped(data, formula, ...)

# S3 method for data.frame
as_ped(data, formula, ...)

# S3 method for nested_fdf
as_ped(data, formula, ...)

# S3 method for list
as_ped(data, formula, ...)

is.ped(x)

Arguments

data
Either an object inheriting from data frame or in case of time-dependent covariates a list of data frames, where the first data frame contains the time-to-event information and static covariates while the second (and potentially further data frames) contain information on time-dependent covariates and the times at which they have been observed.

formula
A two sided formula with a Surv object on the left-hand-side and covariate specification on the right-hand-side (RHS). The RHS can be an extended formula, which specifies how TDCs should be transformed using specials concurrent and cumulative.

...
Further arguments passed to split_data.

x
any R object.

Examples

tumor[1:3, ]

#> # A tibble: 3 x 9
#>    days status charlson_score   age sex    transfusion complications metastases
#>   <dbl>  <int>          <int> <int> <fct>  <fct>       <fct>         <fct>     
#> 1  579.      0              2    58 female yes         no            yes       
#> 2 1192.      0              2    52 male   no          yes           yes       
#> 3  308.      1              2    74 female yes         no            yes       
#> # ... with 1 more variable: resection <fct>

tumor[1:3, ] %>% as_ped(Surv(days, status)~ age + sex, cut = c(0, 500, 1000))

#>   id tstart tend   interval   offset ped_status age    sex
#> 1  1      0  500    (0,500] 6.214608          0  58 female
#> 2  1    500 1000 (500,1000] 4.369448          0  58 female
#> 3  2      0  500    (0,500] 6.214608          0  52   male
#> 4  2    500 1000 (500,1000] 6.214608          0  52   male
#> 5  3      0  500    (0,500] 5.730100          1  74 female

tumor[1:3, ] %>% as_ped(Surv(days, status)~ age + sex)

#>   id tstart tend interval offset ped_status age    sex
#> 1  1      0  308  (0,308] 5.7301          0  58 female
#> 2  2      0  308  (0,308] 5.7301          0  52   male
#> 3  3      0  308  (0,308] 5.7301          1  74 female