dplyr Verbs for ped-Objects

See dplyr documentation of the respective functions for description and examples.

# S3 method for class 'ped'
arrange(.data, ...)

# S3 method for class 'ped'
group_by(.data, ..., .add = FALSE)

# S3 method for class 'ped'
ungroup(x, ...)

# S3 method for class 'ped'
distinct(.data, ..., .keep_all = FALSE)

# S3 method for class 'ped'
filter(.data, ...)

# S3 method for class 'ped'
sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

# S3 method for class 'ped'
sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)

# S3 method for class 'ped'
slice(.data, ...)

# S3 method for class 'ped'
select(.data, ...)

# S3 method for class 'ped'
mutate(.data, ...)

# S3 method for class 'ped'
rename(.data, ...)

# S3 method for class 'ped'
summarise(.data, ...)

# S3 method for class 'ped'
summarize(.data, ...)

# S3 method for class 'ped'
transmute(.data, ...)

# S3 method for class 'ped'
inner_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

# S3 method for class 'ped'
full_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

# S3 method for class 'ped'
left_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

# S3 method for class 'ped'
right_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

Arguments

.data

an object of class ped, see as_ped.

...

see dplyr documentation

x

an object of class ped, see as_ped.

tbl

an object of class ped, see as_ped.

size

<tidy-select> For sample_n(), the number of rows to select. For sample_frac(), the fraction of rows to select. If tbl is grouped, size applies to each group.

replace

Sample with or without replacement?

weight

<tidy-select> Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input. Weights are automatically standardised to sum to 1.

.env

DEPRECATED.

by

A join specification created with join_by(), or a character vector of variables to join by.

If NULL, the default, *_join() will perform a natural join, using all variables in common across x and y. A message lists the variables so that you can check they're correct; suppress the message by supplying by explicitly.

To join on different variables between x and y, use a join_by() specification. For example, join_by(a == b) will match x$a to y$b.

To join by multiple variables, use a join_by() specification with multiple expressions. For example, join_by(a == b, c == d) will match x$a to y$b and x$c to y$d. If the column names are the same between x and y, you can shorten this by listing only the variable names, like join_by(a, c).

join_by() can also be used to perform inequality, rolling, and overlap joins. See the documentation at ?join_by for details on these types of joins.

For simple equality joins, you can alternatively specify a character vector of variable names to join by. For example, by = c("a", "b") joins x$a to y$a and x$b to y$b. If variable names differ between x and y, use a named character vector like by = c("x_a" = "y_a", "x_b" = "y_b").

To perform a cross-join, generating all combinations of x and y, see cross_join().

copy

If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix

If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

Value

a modified ped object (except for do)