Construct a data frame suitable for prediction

Given a data set, returns a data frame type object that can be used as newdata argument in a call to predict and similar functions. The function is particularly usefull in combination with one of the add_* functions, e.g., add_term, add_hazard, etc.

make_newdata(x, ...)

# S3 method for default
make_newdata(x, ...)

# S3 method for ped
make_newdata(x, ...)

# S3 method for fped
make_newdata(x, ...)

Arguments

x
A data frame (or object that inherits from data.frame).

...
Covariate specifications (expressions) that will be evaulated by looking for variables in x. Must be of form z = f(z) where z is a variable in the data set x and f a known function that can be usefully applied to age. See examples below.

Details

Depening on the class of x, mean or modus values will be used for variables not specified in ellipsis. If x is an object that inherits from class ped, useful data set completion will be performed depending on variables specified in ellipsis.

Examples

tumor %>% make_newdata()

#> # A tibble: 1 x 9
#>    days status charlson_score   age sex   transfusion complications metastases
#>   <dbl>  <dbl>          <dbl> <dbl> <fct> <fct>       <fct>         <fct>     
#> 1 1017.  0.483           2.78  62.0 male  no          no            yes       
#> # ... with 1 more variable: resection <fct>

tumor %>% make_newdata(age=c(50))

#> # A tibble: 1 x 9
#>    days status charlson_score   age sex   transfusion complications metastases
#>   <dbl>  <dbl>          <dbl> <dbl> <fct> <fct>       <fct>         <fct>     
#> 1 1017.  0.483           2.78   50. male  no          no            yes       
#> # ... with 1 more variable: resection <fct>

tumor %>% make_newdata(days=seq_range(days, 3), age=c(50, 55))

#> # A tibble: 6 x 9
#>     days status charlson_score   age sex   transfusion complications metastases
#>    <dbl>  <dbl>          <dbl> <dbl> <fct> <fct>       <fct>         <fct>     
#> 1 1.00e0  0.483           2.78   50. male  no          no            yes       
#> 2 1.90e3  0.483           2.78   50. male  no          no            yes       
#> 3 3.81e3  0.483           2.78   50. male  no          no            yes       
#> 4 1.00e0  0.483           2.78   55. male  no          no            yes       
#> 5 1.90e3  0.483           2.78   55. male  no          no            yes       
#> 6 3.81e3  0.483           2.78   55. male  no          no            yes       
#> # ... with 1 more variable: resection <fct>

tumor %>% make_newdata(days=seq_range(days, 3), status=unique(status), age=c(50, 55))

#> # A tibble: 12 x 9
#>       days status charlson_score   age sex   transfusion complications
#>      <dbl>  <int>          <dbl> <dbl> <fct> <fct>       <fct>        
#>  1    1.00      0           2.78   50. male  no          no           
#>  2 1904.        0           2.78   50. male  no          no           
#>  3 3806.        0           2.78   50. male  no          no           
#>  4    1.00      1           2.78   50. male  no          no           
#>  5 1904.        1           2.78   50. male  no          no           
#>  6 3806.        1           2.78   50. male  no          no           
#>  7    1.00      0           2.78   55. male  no          no           
#>  8 1904.        0           2.78   55. male  no          no           
#>  9 3806.        0           2.78   55. male  no          no           
#> 10    1.00      1           2.78   55. male  no          no           
#> 11 1904.        1           2.78   55. male  no          no           
#> 12 3806.        1           2.78   55. male  no          no           
#> # ... with 2 more variables: metastases <fct>, resection <fct>

# mean/modus values of unspecified variables are calculated over whole data
tumor %>% make_newdata(sex=unique(sex))

#> # A tibble: 2 x 9
#>    days status charlson_score   age sex    transfusion complications metastases
#>   <dbl>  <dbl>          <dbl> <dbl> <fct>  <fct>       <fct>         <fct>     
#> 1 1017.  0.483           2.78  62.0 female no          no            yes       
#> 2 1017.  0.483           2.78  62.0 male   no          no            yes       
#> # ... with 1 more variable: resection <fct>

tumor %>% group_by(sex) %>% make_newdata()

#> # A tibble: 2 x 9
#>    days status charlson_score   age sex    transfusion complications metastases
#>   <dbl>  <dbl>          <dbl> <dbl> <fct>  <fct>       <fct>         <fct>     
#> 1 1060.  0.483           2.96  63.3 male   no          no            yes       
#> 2  954.  0.484           2.52  60.1 female no          no            yes       
#> # ... with 1 more variable: resection <fct>

# You can also pass a part of the data sets as data frame to make_newdata
purrr::cross_df(list(days = c(0, 500, 1000), sex = c("male", "female"))) %>%
  make_newdata(x=tumor)

#> # A tibble: 6 x 9
#>    days status charlson_score   age sex    transfusion complications metastases
#>   <dbl>  <dbl>          <dbl> <dbl> <chr>  <fct>       <fct>         <fct>     
#> 1    0.  0.483           2.78  62.0 male   no          no            yes       
#> 2  500.  0.483           2.78  62.0 male   no          no            yes       
#> 3 1000.  0.483           2.78  62.0 male   no          no            yes       
#> 4    0.  0.483           2.78  62.0 female no          no            yes       
#> 5  500.  0.483           2.78  62.0 female no          no            yes       
#> 6 1000.  0.483           2.78  62.0 female no          no            yes       
#> # ... with 1 more variable: resection <fct>

# Examples for PED data
ped <- tumor %>% slice(1:3) %>% as_ped(Surv(days, status)~., cut = c(0, 500, 1000))
ped %>% make_newdata(age=c(50, 55))

#> # A tibble: 2 x 14
#>   tstart  tend intlen interval    id offset ped_status charlson_score   age
#>    <dbl> <dbl>  <dbl> <fct>    <dbl>  <dbl>      <dbl>          <dbl> <dbl>
#> 1     0.  500.   500. (0,500]   1.80   6.21         1.             2.   50.
#> 2     0.  500.   500. (0,500]   1.80   6.21         1.             2.   55.
#> # ... with 5 more variables: sex <fct>, transfusion <fct>, complications <fct>,
#> #   metastases <fct>, resection <fct>

# if time information is specified, other time variables will be specified
# accordingly and offset calculated correctly
ped %>% make_newdata(tend = c(1000), age = c(50, 55))

#>   tstart tend intlen   interval  id   offset ped_status charlson_score age
#> 1    500 1000    500 (500,1000] 1.8 6.214608          1              2  50
#> 2    500 1000    500 (500,1000] 1.8 6.214608          1              2  55
#>      sex transfusion complications metastases resection
#> 1 female         yes            no        yes        no
#> 2 female         yes            no        yes        no

ped %>% make_newdata(tend = unique(tend))

#>   tstart tend intlen   interval  id   offset ped_status charlson_score  age
#> 1      0  500    500    (0,500] 1.8 6.214608          1              2 58.8
#> 2    500 1000    500 (500,1000] 1.8 6.214608          1              2 58.8
#>      sex transfusion complications metastases resection
#> 1 female         yes            no        yes        no
#> 2 female         yes            no        yes        no

ped %>% group_by(sex) %>% make_newdata(tend = unique(tend))

#>   tstart tend intlen   interval       id   offset ped_status charlson_score
#> 1      0  500    500    (0,500] 2.000000 6.214608          1              2
#> 2      0  500    500    (0,500] 1.666667 6.214608          1              2
#> 3    500 1000    500 (500,1000] 2.000000 6.214608          1              2
#> 4    500 1000    500 (500,1000] 1.666667 6.214608          1              2
#>        age    sex transfusion complications metastases resection
#> 1 52.00000   male          no           yes        yes        no
#> 2 63.33333 female         yes            no        yes        no
#> 3 52.00000   male          no           yes        yes        no
#> 4 63.33333 female         yes            no        yes        no