Fit a piece-wise exponential additive model

Basically a wrapper around gam. However, we set family=poisson(), offset=data$offset and method="REML" by default. The first two can not be overriden. The method argument can be specified as usually, but defaults to GCV.cp in gam.

Check if object is of class pamm

pamm(formula, data = list(), method = "REML", ...)

is.pamm(x)

# S3 method for pamm
print(x, ...)

# S3 method for pamm
summary(object, ...)

# S3 method for pamm
plot(x, ...)

Arguments

formula
A GAM formula, or a list of formulae (see formula.gam and also gam.models). These are exactly like the formula for a GLM except that smooth terms, s, te, ti and t2, can be added to the right hand side to specify that the linear predictor depends on smooth functions of predictors (or linear functionals of these).

data
A data frame or list containing the model response variable and covariates required by the formula. By default the variables are taken from environment(formula): typically the environment from which gam is called.

method
The smoothing parameter estimation method. "GCV.Cp" to use GCV for unknown scale parameter and Mallows' Cp/UBRE/AIC for known scale. "GACV.Cp" is equivalent, but using GACV in place of GCV. "REML" for REML estimation, including of unknown scale, "P-REML" for REML estimation, but using a Pearson estimate of the scale. "ML" and "P-ML" are similar, but using maximum likelihood in place of REML. Beyond the exponential family "REML" is the default, and the only other option is "ML".

...
Further arguments passed to gam.

x
Any R object.

object
An object of class pamm as returned by pamm.

See also

gam