Fit a piece-wise exponential additive model

A thin wrapper around gam, however, some arguments are prespecified: family=poisson() and offset=data$offset. These two can not be overwritten. In many cases it will also be advisable to set method="REML".

pamm(formula, data = list(), ..., trafo_args = NULL, engine = "gam")

is.pamm(x)

# S3 method for class 'pamm'
print(x, ...)

# S3 method for class 'pamm'
summary(object, ...)

# S3 method for class 'pamm'
plot(x, ...)

Arguments

formula

A GAM formula, or a list of formulae (see formula.gam and also gam.models). These are exactly like the formula for a GLM except that smooth terms, s, te, ti and t2, can be added to the right hand side to specify that the linear predictor depends on smooth functions of predictors (or linear functionals of these).

data

A data frame or list containing the model response variable and covariates required by the formula. By default the variables are taken from environment(formula): typically the environment from which gam is called.

...

Further arguments passed to engine.

trafo_args

A named list. If data is not in PED format, as_ped will be called internally with arguments provided in trafo_args.

engine

Character name of the function that will be called to fit the model. The intended entries are either "gam" or "bam" (both from package mgcv).

x

Any R object.

object

An object of class pamm as returned by pamm.

See also

Examples

ped <- tumor[1:100, ] %>%
 as_ped(Surv(days, status) ~ complications, cut = seq(0, 3000, by = 50))
pam <- pamm(ped_status ~ s(tend) + complications, data = ped)
summary(pam)
#> 
#> Family: poisson 
#> Link function: log 
#> 
#> Formula:
#> ped_status ~ s(tend) + complications
#> 
#> Parametric coefficients:
#>                  Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)       -7.8839     0.1969 -40.038   <2e-16 ***
#> complicationsyes   0.2453     0.3420   0.717    0.473    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Approximate significance of smooth terms:
#>           edf Ref.df Chi.sq p-value  
#> s(tend) 1.001  1.003  5.456  0.0196 *
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> R-sq.(adj) =  -0.0232   Deviance explained = 1.87%
#> UBRE = -0.79402  Scale est. = 1         n = 1937
## Alternatively
pamm(
 ped_status ~ s(tend) + complications,
 data = tumor[1:100, ],
trafo_args = list(formula = Surv(days, status)~complications))
#> 
#> Family: poisson 
#> Link function: log 
#> 
#> Formula:
#> ped_status ~ s(tend) + complications
#> 
#> Estimated degrees of freedom:
#> 1.75  total = 3.75 
#> 
#> UBRE score: -0.8570005