pammtools: Piece-wise exponential Additive Mixed Modeling tools.

pammtools provides functions and utilities that facilitate fitting Piece-wise Exponential Additive Mixed Models (PAMMs), including data transformation and other convenience functions for pre- and post-processing as well as plotting.

Details

The best way to get an overview of the functionality provided and how to fit PAMMs is to view the vignettes available at https://adibender.github.io/pammtools/articles/. A summary of the vignettes' content is given below:

basics: Introduction to PAMMs and basic modeling.

baseline: Shows how to estimate and visualize baseline model (without covariates) and comparison to respective Cox-PH model.

convenience: Convenience functions for post-processing and plotting PAMMs.

data-transformation: Transforming data into a format suitable to fit PAMMs.

frailty: Specifying "frailty" terms, i.e., random effects for PAMMs.

splines: Specifying spline smooth terms for PAMMs.

strata: Specifying stratified models in which each level of a grouping variable has a different baseline hazard.

tdcovar: Dealing with time-dependent covariates.

tveffects: Specifying time-varying effects.

left-truncation: Estimation for left-truncated data.

competing-risks: Competing risks analysis.

References

Bender, Andreas, Andreas Groll, and Fabian Scheipl. 2018. “A Generalized Additive Model Approach to Time-to-Event Analysis” Statistical Modelling, February. https://doi.org/10.1177/1471082X17748083.

Bender, Andreas, Fabian Scheipl, Wolfgang Hartl, Andrew G. Day, and Helmut Küchenhoff. 2019. “Penalized Estimation of Complex, Non-Linear Exposure-Lag-Response Associations.” Biostatistics 20 (2): 315–31. https://doi.org/10.1093/biostatistics/kxy003.

Bender, Andreas, and Fabian Scheipl. 2018. “pammtools: Piece-Wise Exponential Additive Mixed Modeling Tools.” ArXiv:1806.01042 Stat, June. https://arxiv.org/abs/1806.01042. Ramjith J, Bender A, Roes KCB, Jonker MA. Recurrent events analysis with piece-wise exponential additive mixed models. 2022. Statistical Modelling., 2022

See also

Author

Maintainer: Andreas Bender andreas.bender@stat.uni-muenchen.de (ORCID)

Authors:

Fabian Scheipl fabian.scheipl@stat.uni-muenchen.de (ORCID)

Philipp Kopper philipp.kopper@stat.uni-muenchen.de (ORCID)

Other contributors:

Lukas Burk burk@leibniz-bips.de (ORCID) [contributor]